perf(gatsby-plugin-mdx): prevent babel parse step at sourcing time (#25437)
* perf(gatsby-plugin-mdx): prevent babel parse step at sourcing time

The mdx plugin was doing its default parsing step for every time it got called. At sourcing time it's only called to retrieve the import bindings and for this we can be much faster by manually processing the import statements. So that's what this does.

A very simple flat no-image mdx benchmark cuts down sourcing time in half for this change.

* Support multiple import statements together

* Drop comments