Fixed precondition where it adds perma cache to .json files (#26210)
I've deployed my gatsby site with this web.config with few changes and noticed that `.json` files in other locations are getting cached. Culprit for that was this line,

```xml
                  <add input="{REQUEST_FILENAME}" pattern="(.*\.js)|(.*\.css)" />
```

Since this regex does not detect end of line, it take `.json` files also as `.js` and add permanent cache header to it. Fix for it is add tracking end of line to `IsCachePermanentlyFile` pre-condition like this,

```
                  <add input="{REQUEST_FILENAME}" pattern="((.*\.js)|(.*\.css))$" />
```

Suggesting to fix this for future users like me. 😊